feat: Add the main offer scraping task to tasks/scrape_offer.py
The task script for scraping the main offer from a specified webpage has been added. The script includes the definition of the scrape_main_offer_task function, which fetches the webpage, scrapes specified offer information via selectors, and then leverages a 'write_file' function to store the scraped data as a JSON file.