refactor: web scraping and database insert/update process
Refactored the web scraping and database insert/update process in the utils.py script. The refactoring has improved the code's readability and efficiency by removing unneeded lines of code and simplifying some aspects of the data manipulation process. In particular, it now directly uses the collection saved in the instance instead of repeatedly referencing it through the client and db.